Added /download/{langCode}/{downloadStart} endpoint
This will allow words to be downloaded. Download start is the offset
from which the words will begin. Maximum 5000 words will be sent on a
single response.

If download start is X, words will start from X+1 upto x+5000. Each
response contains a "count" field which indicates how many words are in
the response. This count can be the next offset when calling this
endpoint again