Providing all the varnamd configs through command line flags
Removing the config file because varnamd is primarly used by clints and
it will be difficult to keep configs in multiple places, one in client
side and one in varnamd. So now varnamd is stateless in terms of
configurations. Everything needs to be passed as command line args when
starting varnamd